// RookieEcommerce.AdminSite/src/api/models/CreateProductCommand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateProductCommand
 */
export interface CreateProductCommand {
    /**
     * 
     * @type {string}
     * @memberof CreateProductCommand
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateProductCommand
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CreateProductCommand
     */
    price?: number;
    /**
     * 
     * @type {string}
     * @memberof CreateProductCommand
     */
    categoryId?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateProductCommand
     */
    details?: string | null;
}

/**
 * Check if a given object implements the CreateProductCommand interface.
 */
export function instanceOfCreateProductCommand(value: object): value is CreateProductCommand {
    return true;
}

export function CreateProductCommandFromJSON(json: any): CreateProductCommand {
    return CreateProductCommandFromJSONTyped(json, false);
}

export function CreateProductCommandFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateProductCommand {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'price': json['price'] == null ? undefined : json['price'],
        'categoryId': json['categoryId'] == null ? undefined : json['categoryId'],
        'details': json['details'] == null ? undefined : json['details'],
    };
}

export function CreateProductCommandToJSON(json: any): CreateProductCommand {
    return CreateProductCommandToJSONTyped(json, false);
}

export function CreateProductCommandToJSONTyped(value?: CreateProductCommand | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'description': value['description'],
        'price': value['price'],
        'categoryId': value['categoryId'],
        'details': value['details'],
    };
}